requirements.txt
requests==2.26.0
pyTelegramBotAPI==4.2.2
python-dotenv==0.16.0